requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
black==24.3.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
config==0.5.1
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.0
fastapi-cache2==0.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
orjson==3.9.15
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
time-machine==2.14.1
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
